updated comments and moved to proper directory structure
