--- robotics/robots.argos
@@ -1,98 +1,0 @@
-<?xml version="1.0" ?>
-<argos-configuration>
-
-  <!-- ************************* -->
-  <!-- * Gen/codeeral configuration * -->
-  <!-- ************************* -->
-  <framework>
-    <system threads="0" />
-    <experiment length="0" ticks_per_second="100" />
-  </framework>
-
-  <!-- *************** -->
-  <!-- * Controllers * -->
-  <!-- *************** -->
-  <controllers>
-      <!-- ***************************************************** -->
-    <!-- * This is the Lua controller to associate to robots * -->
-    <!-- ***************************************************** -->
-    <lua_controller id="lua">
-      <!-- Normal actuator/sensor configuration follows -->
-      <actuators>
-        <differential_steering implementation="default" />
-        <range_and_bearing implementation="default" />
-	<leds implementation="default" medium="leds" />
-      </actuators>
-      <sensors>
-        <footbot_proximity implementation="default" show_rays="false" />
-        <positioning implementation="default" show_rays="false" />
-        <footbot_motor_ground implementation="rot_z_only" />
-        <range_and_bearing implementation="medium" medium="rab" />
-        <colored_blob_omnidirectional_camera implementation="rot_z_only" medium="leds" show_rays="true" />
-      </sensors>
-      <params script="controller.lua" />
-    </lua_controller>
-  </controllers>
-
-  <!-- *********************** -->
-  <!-- * Arena configuration * -->
-  <!-- *********************** -->
-  <arena size="4.2, 5.2, 1.2" center="0,0,0.5">
-    <floor id="f" source="image" path="multicolorgrid.png" />
-
-    <box id="bn" size="0.1, 5, 0.2" movable="false">
-      <body position="2,0,0"  orientation="0,0,0" />
-    </box>
-    <box id="bs" size="0.1, 5, 0.2" movable="false">
-      <body position="-2,0,0" orientation="0,0,0" />
-    </box>
-    <box id="be" size="4, 0.1, 0.2" movable="false">
-      <body position="0,-2.5,0" orientation="0,0,0" />
-    </box>
-    <box id="bw" size="4, 0.1, 0.2" movable="false">
-      <body position="0,2.5,0" orientation="0,0,0" />
-    </box>
-
-    <distribute>
-      <position method="uniform" min="-2,-2.5,0" max="2,2.5,0" />
-      <orientation method="gaussian" mean="0,0,0" std_dev="360,0,0" /> 
-      <!-- Set number of robots -->
-      <entity quantity="3" max_trials="1"> 
-        <foot-bot id="fb" 
-                  rab_range = "1">
-          <controller config="lua" />
-        </foot-bot>
-      </entity> 
-    </distribute> 
-
-  </arena>
-
-  <!-- ******************* -->
-  <!-- * Physics engines * -->
-  <!-- ******************* -->
-  <physics_engines>
-    <dynamics2d id="dyn2d" />
-  </physics_engines>
-
-  <!-- ********* -->
-  <!-- * Media * -->
-  <!-- ********* -->
-  <media>
-    <range_and_bearing id="rab" index="grid" grid_size="1,1,1" />
-    <led id="leds" />
-  </media>
-
-  <!-- ****************** -->
-  <!-- * Visualization * -->
-  <!-- ****************** -->
-  <visualization>
-	<qt-opengl lua_editor="false">
-		<camera>
-			<placements>
-				<placement index="0" position="-2.0767,0.015,2.74921" look_at="-1.59254,0.015,1.87423" lens_focal_length="20" />
-			</placements>
-		</camera>
-	</qt-opengl>
-</visualization>
-
-</argos-configuration>
